Move JSRuntimeHelper to a separate file
